Initial Removal of Legacy Views on 6.0.0 beta 2
