--- src/deprecated/TribeEventsListWidget.php
@@ -1,7 +1,0 @@
-<?php
-	_deprecated_file( __FILE__, '3.10', 'Tribe__Events__List_Widget' );
-
-
-	class TribeEventsListWidget extends Tribe__Events__List_Widget {
-
-	}
