--- src/deprecated/Tribe_Template_Factory.php
@@ -1,7 +1,0 @@
-<?php
-	_deprecated_file( __FILE__, '3.10', 'Tribe__Events__Template_Factory' );
-
-
-	class Tribe_Template_Factory extends Tribe__Events__Template_Factory {
-
-	}
