+++ src/functions/template-tags/loop.php
@@ -383,7 +383,7 @@ if ( class_exists( 'Tribe__Events__Main' ) ) {
 	 * @return bool
 	 */
 	function tribe_is_in_main_loop() {
-		return apply_filters( 'tribe_is_main_loop', Tribe__Events__Templates::$isMainLoop );
+		return apply_filters( 'tribe_is_main_loop', false );
 	}
 
 	/**
