+++ src/functions/template-tags/month.php
@@ -13,56 +13,6 @@ if ( ! defined( 'ABSPATH' ) ) {
 if ( class_exists( 'Tribe__Events__Main' ) ) {
 
 	/**
-	 * Display a month
-	 *
-	 * Inline example:
-	 * < code >
-	 * <?php
-	 * // output the events in May 2016 using the full month view template
-	 * tribe_show_month( array( 'eventDate' => '2016-05-01' ) )
-	 * ?>
-	 * </ code >
-	 *
-	 * @param array  $args          query args to pass to the month view
-	 * @param string $template_path template to use, defaults to the full month view
-	 *
-	 **/
-	function tribe_show_month( $args = [], $template_path = 'month/content' ) {
-		if ( tec_events_views_v1_should_display_deprecated_notice() ) {
-			_deprecated_function( __FUNCTION__, '5.13.0', 'On version 6.0.0 this function will be removed. Please refer to <a href="https://evnt.is/v1-removal">https://evnt.is/v1-removal</a> for template customization assistance.' );
-		}
-
-		// temporarily unset the tribe bar params so they don't apply
-		$hold_tribe_bar_args = [];
-		foreach ( $_REQUEST as $key => $value ) {
-			if ( $value && strpos( $key, 'tribe-bar-' ) === 0 ) {
-				$hold_tribe_bar_args[ $key ] = $value;
-				unset( $_REQUEST[ $key ] );
-			}
-		}
-
-		$month_class = new Tribe__Events__Template__Month( $args );
-		$month_class->setup_view();
-
-		do_action( 'tribe_events_before_show_month' );
-
-		tribe_get_template_part( $template_path );
-
-		do_action( 'tribe_events_after_show_month' );
-
-		$month_class->shutdown_view();
-
-		// reinstate the tribe bar params
-		if ( ! empty( $hold_tribe_bar_args ) ) {
-			foreach ( $hold_tribe_bar_args as $key => $value ) {
-				$_REQUEST[ $key ] = $value;
-			}
-		}
-
-		return $month_class->get_events_in_month_ids();
-	}
-
-	/**
 	 * Month view conditional tag
 	 *
 	 * Returns true when on the "real" Month View itself, but not in other secondary instances of the
@@ -86,154 +36,14 @@ if ( class_exists( 'Tribe__Events__Main' ) ) {
 	}
 
 	/**
-	 * Used in the month loop.
-	 * Returns true if there are more calendar days available in the loop.
-	 *
-	 * Example:
-	 * < code >
-	 * <?php
-	 * // loop through the days in the current month query
-	 * while( tribe_events_have_month_days() ) : tribe_events_the_month_day();
-	 *        // do stuff
-	 * endwhile;
-	 * ?>
-	 * </ code >
-	 *
-	 * @return bool
-	 * @see Tribe__Events__Template__Month::have_days()
-	 **/
-	function tribe_events_have_month_days() {
-		return Tribe__Events__Template__Month::have_days();
-	}
-
-	/**
-	 * Used in the month loop.
-	 * Advances the loop pointer to the next day, and sets that day up for use.
-	 *
-	 * Example:
-	 * < code >
-	 * <?php
-	 * // loop through the days in the current month query
-	 * while( tribe_events_have_month_days() ) : tribe_events_the_month_day();
-	 *        // do stuff
-	 * endwhile;
-	 * ?>
-	 * </ code >
-	 *
-	 * @see Tribe__Events__Template__Month::the_day()
-	 **/
-	function tribe_events_the_month_day() {
-		Tribe__Events__Template__Month::the_day();
-	}
-
-	/**
-	 * Used in the month loop.
-	 * Returns the counter for the current week in the month loop
-	 *
-	 * Example:
-	 * < code >
-	 * <?php
-	 * // loop through the days in the current month query
-	 * if( tribe_events_get_current_week == 3 );
-	 *        // do stuff
-	 * endif;
-	 * ?>
-	 * </ code >
-	 *
-	 * @return int
-	 * @see Tribe__Events__Template__Month::get_current_week()
-	 **/
-	function tribe_events_get_current_week() {
-		return Tribe__Events__Template__Month::get_current_week();
-	}
-
-	/**
-	 * Used in the month loop.
-	 * Gets the current day in the month loop
-	 *
-	 * Returned array contains the following elements if the day is in the currently displaying month:
-	 *  'daynum'       => Day of the month (int)
-	 *  'date'         => Complete date (Y-m-d)
-	 *  'events'       => Object containing events on this day (WP_Query)
-	 *  'total_events' => Number of events on this day (int)
-	 *  'view_more'    => Link to the single day (URL)
-	 *
-	 * If the day is part of the previous or next month, the array simply contains:
-	 *    'date' => 'previous' or 'next'
-	 *
-	 * @return array
-	 * @see Tribe__Events__Template__Month::get_current_day()
-	 **/
-	function tribe_events_get_current_month_day() {
-		return apply_filters( 'tribe_events_get_current_month_day', Tribe__Events__Template__Month::get_current_day() );
-	}
-
-	/**
-	 * Used in the month loop.
-	 * Outputs classes for the current month day, including special classes for past / present / future days
-	 *
-	 * @see Tribe__Events__Template__Month::day_classes()
-	 **/
-	function tribe_events_the_month_day_classes() {
-		echo apply_filters( 'tribe_events_the_month_day_class', Tribe__Events__Template__Month::day_classes() );
-	}
-
-	/**
-	 * Used in the month loop.
-	 * Outputs classes for the current single event in the month loop
-	 *
-	 * @see Tribe__Events__Template__Month::event_classes()
-	 **/
-	function tribe_events_the_month_single_event_classes() {
-		echo apply_filters( 'tribe_events_the_month_single_event_classes', Tribe__Events__Template__Month::event_classes() );
-	}
-
-	/**
 	 * Returns whether there are any events in the specific month.
 	 * Independently if there are filters or not.
 	 *
 	 * @return bool
-	 * @see Tribe__Events__Template__Month::get_daily_counts()
 	 * @since 3.1.1
 	 **/
 	function tribe_events_month_has_events() {
 		return apply_filters( 'tribe_events_month_has_events', false );
-	}
-
-	/**
-	 * Returns whether there are any events in the month,
-	 * with the filtered results.
-	 *
-	 * @return bool
-	 *
-	 * @since 4.6.19
-	 * @see Tribe__Events__Template__Month::has_events_filtered()
-	 **/
-	function tribe_events_month_has_events_filtered() {
-		$tribe_month = new Tribe__Events__Template__Month();
-		$has_events  = $tribe_month->has_events_filtered();
-
-		/**
-		 * Filter the result for the check if the month has events after the filters.
-		 *
-		 * @since 4.6.19
-		 */
-		return apply_filters( 'tribe_events_month_has_events_filtered', $has_events );
-	}
-
-
-	/**
-	 * Drop Menu Post Link
-	 *
-	 * Returns the URL where the jump menu sends the month/year request.
-	 *
-	 * @return string URL
-	 */
-	function tribe_get_dropdown_link_prefix() {
-		$tribe_ecp = Tribe__Events__Main::instance();
-		$output    = $tribe_ecp->getLink( 'dropdown' );
-
-		return apply_filters( 'tribe_get_dropdown_link_prefix', $output );
 	}
 
 	/**
